requirements.txt
accelerate==0.27.2
annotated-types==0.6.0
anyio==4.6.2.post1
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contextlib2==21.6.0
Deprecated==1.2.14
diffusers==0.25.1
distro==1.9.0
duckduckgo_search==6.3.2
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
h2==4.1.0
hf_transfer==0.1.6
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
hyperframe==6.0.1
idna==3.10
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2=
langfuse==2.52.1
leptonai==0.15.1
limits==3.9.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
newrelic==9.7.0
numpy==1.26.4
openai==1.13.3
packaging==24.1
pillow==10.2.0
primp==0.6.4
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.20.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
setuptools==68.2.2
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
wheel==0.41.2
wrapt==1.16.0
zipp==3.17.0
